requirements.txt
Flask==3.1.0
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
flask-limiter==3.8.0
